removed old auth and seperated to new struct
