+++ backend/src/layer/data/database/command.rs
@@ -3,10 +3,7 @@ use async_trait::async_trait;
 use sqlx::{Pool, Postgres};
 
 use crate::{
-    interface::persistent_data::{
-        model::{ApiKey, Image},
-        CommandDataAccess, Result,
-    },
+    interface::persistent_data::{model::Image, CommandDataAccess, Result},
     null_error,
     util::{ReportReason, Uuid},
 };
@@ -160,17 +157,6 @@ impl CommandDataAccess for PersistentCommandData {
         .execute(&self.pool)
         .await?;
         Ok(())
-    }
-
-    async fn get_api_keys(&self) -> Result<Vec<ApiKey>> {
-        let keys = sqlx::query_as!(
-            ApiKey,
-            "SELECT api_key as key, description FROM api_key ORDER BY api_key"
-        )
-        .fetch_all(&self.pool)
-        .await?;
-
-        Ok(keys)
     }
 }
 
@@ -444,29 +430,5 @@ mod test {
             .await
             .unwrap()
             .len()
-    }
-
-    #[sqlx::test(fixtures("api_key"))]
-    async fn test_get_api_keys(pool: PgPool) {
-        let command = PersistentCommandData { pool: pool.clone() };
-
-        assert!(command.get_api_keys().await.is_ok());
-        assert_eq!(
-            command.get_api_keys().await.unwrap(),
-            provide_dummy_api_keys()
-        );
-    }
-
-    fn provide_dummy_api_keys() -> Vec<ApiKey> {
-        vec![
-            ApiKey {
-                key: "abc".into(),
-                description: String::new(),
-            },
-            ApiKey {
-                key: "YWpzZGg4MnozNzhkMnppZGFzYXNkMiBzYWZzYSBzPGE5MDk4".into(),
-                description: String::new(),
-            },
-        ]
     }
 }
