+++ backend/src/layer/data/database/mod.rs
@@ -1,4 +1,5 @@
 //! This component is responsible for accesses to the database.
+pub mod auth;
 pub mod command;
 pub mod factory;
 pub mod mealplan_management;
